% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{download_buckettable}
\alias{download_buckettable}
\title{Download buckettable from GNPS}
\usage{
download_buckettable(id, path = "buckettable.csv")
}
\arguments{
\item{id}{GNPS task ID}
}
\value{
The downloaded file is written to disk
}
\description{
Download buckettable from GNPS
}
\examples{

}
