% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{download_edges}
\alias{download_edges}
\title{Download edges from GNPS}
\usage{
download_edges(id, path = "edges.txt")
}
\arguments{
\item{id}{GNPS task ID}
}
\value{
The downloaded file is written to disk
}
\description{
Download edges from GNPS
}
\examples{

}
