% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{download_file}
\alias{download_file}
\title{Download and save a file}
\usage{
download_file(file, path = "NULL")
}
\arguments{
\item{file}{URL to the file}

\item{path}{Path to save the file}
}
\value{
The downloaded file is written to disk
}
\description{
Download and save a file
}
\examples{

}
