% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{download_mgf}
\alias{download_mgf}
\title{Download MGF file from GNPS}
\usage{
download_mgf(id, path = "ions.mgf")
}
\arguments{
\item{id}{GNPS task ID}
}
\value{
The downloaded file is written to disk
}
\description{
Download MGF file from GNPS
}
\examples{

}
