% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{prepare_css}
\alias{prepare_css}
\title{Prepare the CSS matrix from an edges table}
\usage{
prepare_css(edge_path)
}
\arguments{
\item{edge_path}{Path to edges file from GNPS}
}
\value{
CSS matrix
}
\description{
Prepare the CSS matrix from an edges table
}
