% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GNPS_functions.R
\name{read_GNPS_dir}
\alias{read_GNPS_dir}
\title{Read a previously downloaded GNPS dir}
\usage{
read_GNPS_dir(dir = ".")
}
\arguments{
\item{dir}{}
}
\value{
return a list of GNPS files, or -1 if no buckettable file was found
}
\description{
Read a previously downloaded GNPS dir
}
